aliyun-dts-subscribe-python/requirements.txt
avro==1.11.2
kafka-python==2.0.2
lz4==3.1.10
